requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
python-dotenv==1.0.1
Werkzeug==3.0.1
SQLAlchemy==2.0.25
alembic==1.13.1
